database is created and connected through DI

+++ app/src/main/java/com/mk/clucktosurvive/domain/model/ScoreRecord.kt
@@ -2,6 +2,6 @@ package com.mk.clucktosurvive.domain.model
 
 data class ScoreRecord(
     val date: String,
-    val score: Float
+    val score: Int
 
 )
